Fixed 256-token chunked prefill in trace construction pipeline
Rewrite collect_traces.py to use fixed 256-token chunks via mixed_step()
with continuation parameters instead of single-call prefill. Rewrite
build_trace.py batch simulator: replace LIFO preemption with no-preemption
full-sequence pre-allocation and fixed-chunk scheduling. Switch router
input to float32 for trace recording precision. Add 8 new test classes
for chunk boundary invariants and token budget capping. Add
validate_batched_trace.py for trace consistency checks.

+++ src/MoE/expert_offload_engine.py
@@ -244,7 +244,7 @@ class ExpertOffloadEngine:
             self._router_inputs.append({
                 'step': self._current_step,
                 'layer': layer,
-                'hidden': router_input_buf[:n_tokens].half().cpu().numpy(),
+                'hidden': router_input_buf[:n_tokens].float().cpu().numpy(),
             })
 
         # Load missing experts (before stage4b graph replay)
@@ -414,7 +414,7 @@ class ExpertOffloadEngine:
             self._router_inputs.append({
                 'step': self._current_step,
                 'layer': layer,
-                'hidden': router_input_buf[:n_tokens].half().cpu().numpy(),
+                'hidden': router_input_buf[:n_tokens].float().cpu().numpy(),
             })
 
         # Step 4: identify what's still missing after pre-attn prefetch
